feat: configurable fn to retrieve spans and allowed qualifiers in
