+++ edsnlp/pipelines/trainable/span_qualifier/__init__.py
@@ -1,0 +1,1 @@
+from .factory import create_candidate_getter, create_component, create_scorer
