+++ tests/pipelines/trainable/test_span_classifier.py
@@ -3,7 +3,7 @@ from pytest import fixture, mark
 from spacy.tokens import Span
 from spacy.training import Corpus, Example
 
-from edsnlp.pipelines.trainable.span_qualifier import SPAN_CLASSIFIER_DEFAULTS
+from edsnlp.pipelines.trainable.span_qualifier.factory import SPAN_QUALIFIER_DEFAULTS
 from edsnlp.utils.training import make_spacy_corpus_config, train
 
 if not Span.has_extension("label"):
@@ -24,12 +24,7 @@ def gold(blank_nlp):
 
     doc1.spans["sc"] = [
         # drug = "folfox"
-        Span(
-            doc1,
-            4,
-            5,
-            "drug",
-        ),
+        Span(doc1, 4, 5, "drug"),
         # event = "Arret"
         Span(doc1, 0, 1, "event"),
         # criteria = "si"
@@ -80,12 +75,12 @@ def test_span_qualifier_label_training(gold, tmp_path):
     nlp.add_pipe(
         "span_qualifier",
         config={
-            **SPAN_CLASSIFIER_DEFAULTS,
+            **SPAN_QUALIFIER_DEFAULTS,
             "qualifiers": ("label_",),
-            "on_ents": False,
-            "on_span_groups": True,
+            "from_ents": False,
+            "from_span_groups": True,
             "model": {
-                **SPAN_CLASSIFIER_DEFAULTS["model"],
+                **SPAN_QUALIFIER_DEFAULTS["model"],
             },
         },
     )
@@ -120,7 +115,7 @@ def test_span_qualifier_label_training(gold, tmp_path):
     pred = nlp(pred)
     scores = nlp.pipeline[-1][1].score([Example(pred, gold[0])])
     assert [span.label_ for span in pred.spans["sc"]] == ["drug", "event", "criteria"]
-    assert scores["accuracy"] == 1.0
+    assert scores["qual_f"] == 1.0
 
 
 @mark.parametrize("lang", ["eds"], indirect=True)
@@ -131,16 +126,15 @@ def test_span_qualifier_constrained_training(gold, tmp_path):
     nlp.add_pipe(
         "span_qualifier",
         config={
-            **SPAN_CLASSIFIER_DEFAULTS,
-            "qualifiers": ("_.test_negated", "_.event_type"),
-            "on_ents": True,
-            "on_span_groups": {"sc": ["event", "drug", "criteria"]},
-            "model": {
-                **SPAN_CLASSIFIER_DEFAULTS["model"],
+            **SPAN_QUALIFIER_DEFAULTS,
+            "candidate_getter": {
+                "@misc": "eds.candidate_span_qualifier_getter",
+                "qualifiers": ("_.test_negated", "_.event_type"),
+                "label_constraints": {"_.event_type": ("event",)},
+                "from_ents": False,
+                "from_span_groups": ("sc",),
             },
-            "ner_constraints": {
-                "_.event_type": ("event",),
-            },
+            "model": SPAN_QUALIFIER_DEFAULTS["model"],
         },
     )
 
@@ -156,7 +150,6 @@ def test_span_qualifier_constrained_training(gold, tmp_path):
             **{
                 "training.max_steps": 5,
                 "training.eval_frequency": 5,
-                # "training.optimizer.learn_rate": 0,
             },
         ),
     )
@@ -173,4 +166,4 @@ def test_span_qualifier_constrained_training(gold, tmp_path):
     scores = nlp.pipeline[-1][1].score([Example(pred, gold[0])])
     assert [s._.test_negated for s in pred.spans["sc"]] == [False, True, False]
     assert [s._.event_type for s in pred.spans["sc"]] == [None, "stop", None]
-    assert scores["accuracy"] == 1.0
+    assert scores["qual_f"] == 1.0
